MQT-6384 revert SRTP to 1.6

+++ libs/srtp/crypto/test/datatypes_driver.c
@@ -9,7 +9,7 @@
 
 /*
  *	
- * Copyright (c) 2001-2017, Cisco Systems, Inc.
+ * Copyright (c) 2001-2006, Cisco Systems, Inc.
  * All rights reserved.
  * 
  * Redistribution and use in source and binary forms, with or without
@@ -51,7 +51,6 @@
 #include <stdio.h>            /* for printf() */
 #include <string.h>           /* for strlen() */
 #include "datatypes.h"
-#include "util.h"
 
 void
 byte_order(void);
@@ -227,7 +226,7 @@ test_bswap(void) {
   uint32_t x = 0x11223344;
   uint64_t y = 0x1122334455667788LL;
 
-  printf("before: %0x\nafter:  %0x\n", x, (unsigned int)be32_to_cpu(x));
+  printf("before: %0x\nafter:  %0x\n", x, be32_to_cpu(x));
   printf("before: %0llx\nafter:  %0llx\n", (unsigned long long)y,
 	 (unsigned long long)be64_to_cpu(y));
